refactor(writer): split BIG vs BTI write paths out of writer/mod.rs (#1128)
Behavior-preserving module split (epic #1116). Extracts two private
sibling submodules from the SSTableWriter orchestrator, leaving mod.rs
as the entry points + struct + re-exports:

- writer/bti_state.rs: BTI (`da`) deferred-payload state
  (PendingBtiPartition/PendingRowIndex) + the BTI write path —
  queue_bti_partition (per-partition trie payload deferral) and
  write_bti_components (Rows.db/Partitions.db serialization). Isolates
  every BTI-only branch out of write_partition/finish.
- writer/finish.rs: format-aware finalization (finish()) plus the shared
  component-path / bloom-filter / CRC32 helpers it owns.

No semantic changes: Data.db and all component bytes are unchanged for
both BIG and BTI. write_bti_components takes the deferred state by value
so finish() can call it after summary_writer.finish() partially moves
self. Tests for the writer lifecycle stay next to the struct in mod.rs
(they exercise the public entry points + the private `stats` field);
count preserved (303 lib tests pass).

mod.rs: 2705 -> 2288 lines (non-test source ~922; tests ~1365).
Does not touch data_writer/stats_writer (#1118/#1129 own those splits).
